fix: compilation errors

+++ src/ros/rover/arm/arm_kinematics/arm_kinematics/include/arm_kinematics/forward/utilities/eigen_forward_kinematics_plugin.hpp
@@ -53,7 +53,7 @@ public:
   private:
     ComputeFrameTree tree_;
     JointMap joint_map_;
-    std::vector<double> mapped_joint_states_{};
+    std::vector<float> mapped_joint_states_{};
   };
 
   /**
